Use day-draft hook
