--- src/lib/create-event-shortcut.ts
@@ -1,1 +1,0 @@
-export const CREATE_EVENT_ON_ACTIVE_DAY = "rencal:create-event-on-active-day"
